% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bioc2mlr-package.R
\docType{package}
\name{Bioc2mlr-package}
\alias{Bioc2mlr}
\alias{Bioc2mlr-package}
\title{Bioc2mlr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
