% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_SE_to_taskFunc.R
\name{Fun_SE_to_taskFunc}
\alias{Fun_SE_to_taskFunc}
\title{Convert SE to mlr task}
\usage{
Fun_SE_to_taskFunc(SE_obj, param.Y.name, param.covariates,
  param_positive_y_level, task_return_format, task_type)
}
\arguments{
\item{SE_obj}{SE class}

\item{param.Y.name}{Vector of dependent variable name}

\item{param.covariates}{Vector of coaraiate variable(s) name}

\item{param_positive_y_level}{if ClassifTask, value (character or numeric) to be considered as the positive factor outcome}
}
\value{
mlr's \code{ClassifTask} or \code{RegrTask}
}
\description{
Convert SE to mlr task
}
\details{
SE with multiple sub-assays, will include each sub-assay separately.

Either ClassifTask or RegrTask will be returned, based on the type of the param.Y.name variable
}
\examples{
data(airway, package="airway") # RangedSummarizedExperiment
airway
Fun_SE_to_taskFunc(airway, param.Y.name = 'dex', param.covariates = c('avgLength'), param_positive_y_level = 'trt')

}
