% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaCPO.R
\name{limmaCPO}
\alias{limmaCPO}
\title{Add together two numbers}
\usage{
limmaCPO(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\details{
In
}
\examples{
add(1, 1)
add(10, 1)
}
