% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_MAE.cohort2.sam.mine.R
\docType{data}
\name{MAE.cohort2.sam.mine}
\alias{MAE.cohort2.sam.mine}
\title{Dataset, Multi assay data in mlr's task format}
\format{mlr's task
\describe{
  \item{MAE.cohort2.sam.mine}{task}
  ...
}}
\source{
JDRF multi center efforts
}
\usage{
MAE.cohort2.sam.mine
}
\description{
mlr's task (NO functionals implemented YET), therefore, use sep to at analyte name to exatract assay annotation
}
\keyword{datasets}
