% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_1_helpers.R
\name{post_processing_helpers}
\alias{post_processing_helpers}
\alias{whole_run_function}
\alias{whole_run_MLR_function}
\title{Helper post_processing functions}
\usage{
whole_run_function(task, lrn, lambda_opt, index)

whole_run_MLR_function(task, lrn, adjust_alt_lambda)
}
\arguments{
\item{all_kind}{all kind...}
}
\value{
all kind...
}
\description{
Helper post_processing functions
}
\details{
TBA
}
