% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_Filter_1_helpers.R
\name{Wrapper_Filter_1_helpers}
\alias{Wrapper_Filter_1_helpers}
\alias{Func_update_args_univ_clusters}
\alias{getTaskSubjectId}
\alias{Func_extract_feature_name}
\alias{Filter_M_Limma_eBayes_gene_REV_function}
\alias{Filter_M_glm_gene_REV_function}
\alias{Fun_Load_module_sets}
\alias{F_assay_i_module_row_index}
\alias{Filter_M_Limma_camera_Module_REV_function}
\alias{Func_task_to_Passed_Features_DF_MaG}
\alias{Func_Collapse_Data_by_module_names}
\alias{cvglmnet_alternative_lambda.min_function}
\title{Helper functions}
\usage{
Func_update_args_univ_clusters(lrn, args_vec, lrn.id)

getTaskSubjectId(Task)

Func_extract_feature_name(DF_Assay_feature_names, Assay.Analyte.sep)

Filter_M_Limma_eBayes_gene_REV_function(DF.i, y.either,
  Covaraites_DF_const)

Filter_M_glm_gene_REV_function(DF.i, y.either, Covaraites_DF_const)

Fun_Load_module_sets(mod_all.sets.tib)

F_assay_i_module_row_index(DF_i, modules_sets_combined_l_F)

Filter_M_Limma_camera_Module_REV_function(DF_i, y.either,
  Covaraites_DF_const_F, Set_i, assay_i)

Func_task_to_Passed_Features_DF_MaG(task.multi, args)

Func_Collapse_Data_by_module_names(DF_gene, Train_feature_names)

cvglmnet_alternative_lambda.min_function(cv_fit,
  param.LASSO.n.features.arbitrary)
}
\arguments{
\item{all_kind}{all kind...}
}
\value{
all kind...
}
\description{
Helper functions
}
\details{
TBA

when cv.glmnet lambda.min return a null model (intercept only), alternatively return 'previous'
value of lambda that reults \code {param.LASSO.n.features.arbitrary} features
}
