% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_Filter_2_train_predict.R
\name{Wrapper_Filter_2_train_predict}
\alias{Wrapper_Filter_2_train_predict}
\alias{F_PreProc_1_Univ_Only_Train_MaG}
\alias{F_PreProc_3_UnivClust_Train_MaG}
\alias{F_PreProc13_BOTH_Predict_MaG}
\title{Customized train and predict functions for a customized preprocessing wrapper}
\usage{
F_PreProc_1_Univ_Only_Train_MaG(data, target,
  args = list(param.Univ.filt.top.n.features =
  param.Univ.filt.top.n.features, parame.gene.or.module =
  parame.gene.or.module, param.LASSO.n.features.arbitrary =
  param.LASSO.n.features.arbitrary))

F_PreProc_3_UnivClust_Train_MaG(data, target,
  args = list(param.Univ.filt.top.n.features =
  param.Univ.filt.top.n.features, param.UnivClustRankTopN =
  param.UnivClustRankTopN, param.cluster_method_KH =
  param.cluster_method_KH, param.corrplot.n.clusters.k =
  param.corrplot.n.clusters.k, param.corrplot.n.clusters.h =
  param.corrplot.n.clusters.h, parame.gene.or.module =
  parame.gene.or.module, param.LASSO.n.features.arbitrary =
  param.LASSO.n.features.arbitrary))

F_PreProc13_BOTH_Predict_MaG(data, target, args, control)
}
\arguments{
\item{data}{data}

\item{target}{target (character)}

\item{args}{args (list)}

\item{control}{control (predict only)}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Customized train and predict functions for a customized preprocessing wrapper
}
\details{
In
}
\examples{
F_PreProc_1_Univ_Only_Train_MaG(data, target, args)
F_PreProc_3_UnivClust_Train_MaG(data, target, args)
F_PreProc13_BOTH_Predict_MaG(data, target, args, control)
}
