% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper_Filter_3_makePrep_MaG.R
\name{Wrapper_Filter_3_makePrep_MaG}
\alias{Wrapper_Filter_3_makePrep_MaG}
\alias{Fun_lrn_univ_only_makePrep_MaG}
\alias{Fun_lrn_univ_Clusters_All_makePrep_MaG}
\title{Customized preprocessing wrapper. wrappers around makePreprocWrapper: Univariate only; Univariate + Cluster}
\usage{
Fun_lrn_univ_only_makePrep_MaG(learner, train_F, Predict_F,
  param.Univ.filt.top.n.features, parame.gene.or.module,
  param.LASSO.n.features.arbitrary)

Fun_lrn_univ_Clusters_All_makePrep_MaG(learner, train_F, Predict_F,
  param.Univ.filt.top.n.features, param.UnivClustRankTopN,
  param.cluster_method_KH, param.corrplot.n.clusters.k,
  param.corrplot.n.clusters.h, parame.gene.or.module,
  param.LASSO.n.features.arbitrary, param.assay.type.vec)
}
\arguments{
\item{learner}{A learner}

\item{train_F}{train_F}

\item{Predict_F}{Predict_F}

\item{param.Univ.filt.top.n.features}{param.Univ.filt.top.n.features}

\item{parame.gene.or.module}{parame.gene.or.module}

\item{param.LASSO.n.features.arbitrary}{param.LASSO.n.features.arbitrary}

\item{param.UnivClustRankTopN}{param.UnivClustRankTopN}

\item{param.cluster_method_KH}{param.cluster_method_KH}

\item{param.corrplot.n.clusters.k}{param.corrplot.n.clusters.k}

\item{param.corrplot.n.clusters.h}{param.corrplot.n.clusters.h}

\item{param.assay.type.vec}{param.assay.type.vec}

\item{parame.gene.or.module}{parame.gene.or.module}

\item{param.LASSO.n.features.arbitrary}{param.LASSO.n.features.arbitrary}
}
\value{
\code{learner} custom preprocessing wrapper with makePreprocWrapper() of class 'PreprocWrapper'
}
\description{
Customized preprocessing wrapper. wrappers around makePreprocWrapper: Univariate only; Univariate + Cluster
}
\details{
TBA
}
\examples{
My_learner<-lrn
F_PreProc_1_Univ_Only_Train_MaG<-A
F_PreProc_3_UnivClust_Train_MaG<-B
Fun_lrn_univ_only_makePrep_MaG(        learner = My_learner, train_F = F_PreProc_1_Univ_Only_Train_MaG, Predict_F = F_PreProc13_BOTH_Predict_MaG, param.Univ.filt.top.n.features = 10, parame.gene.or.module = 'gene', param.LASSO.n.features.arbitrary = '5')
Fun_lrn_univ_Clusters_All_makePrep_MaG(learner = My_learner, train_F = F_PreProc_3_UnivClust_Train_MaG, Predict_F = F_PreProc13_BOTH_Predict_MaG, param.Univ.filt.top.n.features, param.UnivClustRankTopN, param.cluster_method_KH, param.corrplot.n.clusters.k, param.corrplot.n.clusters.h, parame.gene.or.module, param.LASSO.n.features.arbitrary, param.assay.type.vec)
}
\references{
https://pat-s.github.io/mlr/articles/tutorial/devel/preproc.html#creating-the-preprocessing-wrapper
}
