% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionals.R
\name{functional_helpers}
\alias{functional_helpers}
\alias{functional_to_NonFunctional_task_function}
\alias{next_F}
\alias{task_classif_to_Unsupervised}
\title{Functionals}
\usage{
functional_to_NonFunctional_task_function(task_functional)

next_F(task)

task_classif_to_Unsupervised(task_non_functional)
}
\arguments{
\item{task}{task}
}
\value{
\code{task}
}
\description{
Functionals
}
\details{
Convert functional task, to non-functional task
}
\examples{
fd.features = list("UVVIS" = 3:136, "NIR" = 137:367)
fdf = makeFunctionalData(df, fd.features = fd.features)
tsk_functional = makeRegrTask("fuelsubset", data = fdf, target = "heatan")
tsk_functional
tsk_non_functional<-functional_to_NonFunctional_task_function(tsk_functional)
}
